humpy: Add support for keyword humpification
Some libraries (I'm looking at you, PyQT!) use camelCase for functions,
but CapitalLettersAtFront for variables and types. Since we want to
access both as snake-names, we need to convert the latter to the proper
dialect of camel-case.

To do this, we will use keywords: if the #@ reader macro sees a keyword,
it will capitalize the first letter too, otherwise it will behave as it
did before.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>